+++ packages/charts/src/internal/ChartContainer.tsx
@@ -1,9 +1,10 @@
-import { createComponentStyles } from '@ui5/webcomponents-react-base/lib/createComponentStyles';
+import { createUseStyles } from 'react-jss';
 import { ThemingParameters } from '@ui5/webcomponents-react-base/lib/ThemingParameters';
 import { usePassThroughHtmlProps } from '@ui5/webcomponents-react-base/lib/usePassThroughHtmlProps';
 import { CommonProps } from '@ui5/webcomponents-react/interfaces/CommonProps';
+import { Label } from '@ui5/webcomponents-react/lib/Label';
 import { Loader } from '@ui5/webcomponents-react/lib/Loader';
-import React, { ComponentType, CSSProperties, FC, forwardRef, ReactElement, Ref, useMemo } from 'react';
+import React, { ComponentType, CSSProperties, FC, forwardRef, ReactElement, ReactNode, Ref, useMemo } from 'react';
 import { ResponsiveContainer } from 'recharts';
 
 export interface ContainerProps extends CommonProps {
@@ -39,7 +40,26 @@ const chartContainerStyles = {
   }
 };
 
-const useStyles = createComponentStyles(chartContainerStyles, { name: 'ChartContainer' });
+const useStyles = createUseStyles(chartContainerStyles, { name: 'ChartContainer' });
+
+class ErrorBoundary extends React.Component<{ children: ReactNode }, { errorCount: number }> {
+  state = {
+    errorCount: 0
+  };
+
+  componentDidCatch() {
+    if (this.state.errorCount < 3) {
+      this.setState((old) => ({ ...old, errorCount: old.errorCount + 1 }));
+    }
+  }
+
+  render() {
+    if (this.state.errorCount >= 3) {
+      return <Label>Sorry, something went wrong while rendering this chart!</Label>;
+    }
+    return this.props.children;
+  }
+}
 
 const ChartContainer: FC<ContainerProps> = forwardRef((props: ContainerProps, ref: Ref<any>) => {
   const { Placeholder, loading = false, dataset, style, className, tooltip, slot, children, resizeDebounce } = props;
@@ -64,11 +84,11 @@ const ChartContainer: FC<ContainerProps> = forwardRef((props: ContainerProps, re
       {dataset?.length > 0 ? (
         <>
           {loading && <Loader style={loaderStyles} />}
-          {
+          <ErrorBoundary>
             <ResponsiveContainer debounce={resizeDebounce} {...__testingProps__}>
               {children}
             </ResponsiveContainer>
-          }
+          </ErrorBoundary>
         </>
       ) : (
         <Placeholder />