+++ packages/main/src/internal/withWebComponent.tsx
@@ -1,4 +1,5 @@
 import { getEffectiveScopingSuffixForTag } from '@ui5/webcomponents-base/dist/CustomElementsScope';
+import { isIE } from '@ui5/webcomponents-react-base/lib/Device';
 import { useConsolidatedRef } from '@ui5/webcomponents-react-base/lib/useConsolidatedRef';
 import React, {
   Children,
@@ -23,6 +24,8 @@ const kebabToCamelCase = (str: string) => str.replace(/([-_]\w)/g, (g) => g[1].t
 const createEventPropName = (eventName) => `on${capitalizeFirstLetter(kebabToCamelCase(eventName))}`;
 
 type EventHandler = (event: CustomEvent<unknown>) => void;
+
+const staticSlotStyle = { display: 'contents' };
 
 export interface WithWebComponentPropTypes extends CommonProps {
   ref?: Ref<any>;
@@ -58,23 +61,35 @@ export const withWebComponent = <T extends Record<string, any>>(
       return acc;
     }, {});
 
-    const slots = slotProperties.reduce((acc, name) => {
-      const slotValue = rest[name] as ReactElement;
-      if (slotValue) {
-        return [
-          ...acc,
-          ...Children.map(
-            slotValue?.type === React.Fragment ? slotValue.props?.children : slotValue,
-            (item: ReactElement, index) =>
-              cloneElement(item, {
-                key: `${name}-${index}`,
-                slot: name
-              })
-          )
-        ];
-      }
-      return acc;
-    }, []);
+    const slots = isIE()
+      ? slotProperties.reduce((acc, name) => {
+          const slotValue = rest[name] as ReactElement;
+          if (slotValue) {
+            return [
+              ...acc,
+              ...Children.toArray(slotValue?.type === React.Fragment ? slotValue.props?.children : slotValue)
+                .filter(Boolean)
+                .map((item: ReactElement, index) =>
+                  cloneElement(item, {
+                    key: `${name}-${index}`,
+                    slot: name
+                  })
+                )
+            ];
+          }
+          return acc;
+        }, [])
+      : slotProperties
+          .map((item) => {
+            const slotValue = rest[item] as ReactElement;
+            if (!slotValue) return false;
+            return (
+              <div slot={item} style={staticSlotStyle} key={item}>
+                {slotValue}
+              </div>
+            );
+          })
+          .filter(Boolean);
 
     // event binding
     useEffect(